refactor: move `Card` from `shared` layer to `features`
